100% DAO implementation

+++ src/JLMS/dao/HoldDao.java
@@ -6,9 +6,9 @@ import java.sql.Date;
 import java.util.List;
 
 public interface HoldDao extends GenericDao  {
-    Hold getByBookPatron(int book_id, String patron_username) throws Exception;
+    Hold getByBookPatron(long book_id, String patron_username) throws Exception;
 
     List<Hold> getByPatron_username(String patron_username) throws Exception;
 
-    List<Hold> getByBook_id(int book_id) throws Exception;
+    List<Hold> getByBook_id(long book_id) throws Exception;
 }
